// apps/web/src/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4";
  };
  public: {
    Tables: {
// ...
      cases: {
        Row: {
          created_at: string | null;
          entity_extraction: Json | null;
          external_id: string | null;
          extreme_case_check: Json | null;
          id: string;
          metadata: Json | null;
          scheduled_at: string | null;
          source: string | null;
          status: Database["public"]["Enums"]["CaseStatus"] | null;
          type: Database["public"]["Enums"]["CaseType"] | null;
          updated_at: string | null;
          user_id: string | null;
          visibility: Database["public"]["Enums"]["CaseVisibility"];
        };
        Insert: {
          created_at?: string | null;
          entity_extraction?: Json | null;
          external_id?: string | null;
          extreme_case_check?: Json | null;
          id?: string;
          metadata?: Json | null;
          scheduled_at?: string | null;
          source?: string | null;
          status?: Database["public"]["Enums"]["CaseStatus"] | null;
          type?: Database["public"]["Enums"]["CaseType"] | null;
          updated_at?: string | null;
          user_id?: string | null;
          visibility?: Database["public"]["Enums"]["CaseVisibility"];
        };
        Update: {
          created_at?: string | null;
          entity_extraction?: Json | null;
          external_id?: string | null;
          extreme_case_check?: Json | null;
          id?: string;
          metadata?: Json | null;
          scheduled_at?: string | null;
          source?: string | null;
          status?: Database["public"]["Enums"]["CaseStatus"] | null;
          type?: Database["public"]["Enums"]["CaseType"] | null;
          updated_at?: string | null;
          user_id?: string | null;
          visibility?: Database["public"]["Enums"]["CaseVisibility"];
        };
        Relationships: [
          {
            foreignKeyName: "cases_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      patients: {
        Row: {
          breed: string | null;
          case_id: string | null;
          created_at: string;
          date_of_birth: string | null;
          external_id: string | null;
          id: string;
          metadata: Json | null;
          name: string;
          owner_email: string | null;
          owner_name: string | null;
          owner_phone: string | null;
          sex: string | null;
          source: string | null;
          species: string | null;
          updated_at: string | null;
          user_id: string | null;
          weight_kg: number | null;
        };
        Insert: {
          breed?: string | null;
          case_id?: string | null;
          created_at?: string;
          date_of_birth?: string | null;
          external_id?: string | null;
          id?: string;
          metadata?: Json | null;
          name: string;
          owner_email?: string | null;
          owner_name?: string | null;
          owner_phone?: string | null;
          sex?: string | null;
          source?: string | null;
          species?: string | null;
          updated_at?: string | null;
          user_id?: string | null;
          weight_kg?: number | null;
        };
        Update: {
          breed?: string | null;
          case_id?: string | null;
          created_at?: string;
          date_of_birth?: string | null;
          external_id?: string | null;
          id?: string;
          metadata?: Json | null;
          name?: string;
          owner_email?: string | null;
          owner_name?: string | null;
          owner_phone?: string | null;
          sex?: string | null;
          source?: string | null;
          species?: string | null;
          updated_at?: string | null;
          user_id?: string | null;
          weight_kg?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "patient_case_id_fkey";
            columns: ["case_id"];
            isOneToOne: false;
            referencedRelation: "cases";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "patients_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;